Move all main selectors to the main state management domain
