+++ src/stores/main/selectors.ts
@@ -3,7 +3,12 @@ import { isJwtExpired } from 'lib/helpers';
 import { IState } from 'reducers/types';
 import { AnyAction } from 'redux';
 import { ThunkDispatch } from 'redux-thunk';
-import { AuthorizationTypes, StatusCodes } from 'shared/constants';
+import {
+  AuthorizationTypes,
+  Constants,
+  Providers,
+  StatusCodes,
+} from 'shared/constants';
 import { PropertiesOf } from 'shared/types';
 import { auth0Login } from 'stores/main/actions';
 
@@ -13,6 +18,10 @@ export interface ISSOError {
 
 export function getUserIsAdmin(state: IState) {
   return state.main.loggedInUser?.isAdmin ?? false;
+}
+
+export function getProvider(state: IState): PropertiesOf<typeof Providers> {
+  return state.main.info.general.provider;
 }
 
 /**
@@ -57,3 +66,53 @@ export const selectAuthToken = (
     }
   };
 };
+
+export function getMinHAMastersVersion(state: IState): string {
+  const provider = getProvider(state);
+  let releaseVersion = '';
+
+  switch (provider) {
+    case Providers.AWS:
+      releaseVersion = Constants.AWS_HA_MASTERS_VERSION;
+      break;
+  }
+
+  if (state.main.info.features?.ha_masters?.release_version_minimum) {
+    releaseVersion =
+      state.main.info.features.ha_masters.release_version_minimum;
+  }
+
+  return releaseVersion;
+}
+
+export function getFirstNodePoolsRelease(state: IState): string {
+  const provider = getProvider(state);
+  let releaseVersion = '';
+
+  switch (provider) {
+    case Providers.AWS:
+      releaseVersion = Constants.AWS_V5_VERSION;
+      break;
+
+    case Providers.AZURE:
+      releaseVersion = Constants.AZURE_V5_VERSION;
+      break;
+  }
+
+  if (state.main.info.features?.nodepools) {
+    releaseVersion = state.main.info.features.nodepools.release_version_minimum;
+  }
+
+  return releaseVersion;
+}
+
+export function getAllowedInstanceTypeNames(state: IState): string[] {
+  switch (state.main.info.general.provider) {
+    case Providers.AWS:
+      return state.main.info.workers.instance_type?.options ?? [];
+    case Providers.AZURE:
+      return state.main.info.workers.vm_size?.options ?? [];
+    default:
+      return [];
+  }
+}
